Remove the auto loading of bundler
This will make sure that wars don't have implicit behavior if bundler *can* be
loaded. This is undesirable since bundler might somehow be findable by Ruby,
but might not be part of the .war package.

If we load bundler, and the .war doesn't have a Gemfile then exceptions are
raised and sadness is had

Fixes #4